chore(deps)!: require bevy version 0.6 (#15)
BREAKING CHANGE: the cargo feature `warnings` is removed, as it is no longer possible to add the animation as component by mistake

+++ src/animation.rs
@@ -29,7 +29,7 @@ pub enum AnimationMode {
 #[derive(Debug, Copy, Clone, Default)]
 pub struct Frame {
     /// Index in the sprite atlas
-    pub index: u32,
+    pub index: usize,
     /// How long should the frame be displayed
     pub duration: Duration,
 }
@@ -48,7 +48,7 @@ impl SpriteSheetAnimation {
     ///
     /// For more granular configuration, see [`from_frames`](SpriteSheetAnimation::from_frames)
     #[must_use]
-    pub fn from_range(index_range: RangeInclusive<u32>, frame_duration: Duration) -> Self {
+    pub fn from_range(index_range: RangeInclusive<usize>, frame_duration: Duration) -> Self {
         Self::from_frames(
             index_range
                 .map(|index| Frame::new(index, frame_duration))
@@ -86,7 +86,7 @@ impl Frame {
     /// Create a new animation frame
     #[inline]
     #[must_use]
-    pub fn new(index: u32, duration: Duration) -> Self {
+    pub fn new(index: usize, duration: Duration) -> Self {
         Self { index, duration }
     }
 }
